refactor/remove thread and add api to bot
