--- base/__init__.py
@@ -1,6 +1,0 @@
-from .settings import *
-from .event_handlers import *
-
-from bot_features import *
-
-# from api import *
