+++ bot/__init__.py
@@ -1,0 +1,3 @@
+from .settings import *
+from .event_handlers import *
+from .features import *
