+++ bot/features/__init__.py
@@ -1,0 +1,12 @@
+## Event and schedule
+from .schedule import *
+from .guild_event import *
+
+## Features
+from .security import *
+from .errands import *
+
+from .confession import *
+from .create_vc import *
+
+# from .manage import *
