+++ bot/features/confession/__init__.py
@@ -1,0 +1,9 @@
+from .confession import *
+
+
+from bot import bot
+
+
+@bot.listen()
+async def on_ready():
+    print("5.Confession ready")
