+++ bot/features/manage/__init__.py
@@ -1,0 +1,6 @@
+from bot import bot
+
+
+@bot.listen()
+async def on_ready():
+    print("10.Manage staff ready")
