--- bot_features/confession/__init__.py
@@ -1,9 +1,0 @@
-from .confession import *
-
-
-from base import bot
-
-
-@bot.listen()
-async def on_ready():
-    print("5.Confession ready")
