--- bot_features/errands/__init__.py
@@ -1,11 +1,0 @@
-from .color import *
-from .topic import *
-from .game_center import *
-from .welcome import *
-
-from base import bot
-
-
-@bot.listen()
-async def on_ready():
-    print("3.Errand ready")
