--- bot_features/manage/__init__.py
@@ -1,5 +1,0 @@
-from base import bot
-
-@bot.listen()
-async def on_ready():
-    print('10.Manage staff ready')
