--- bot_features/schedule/__init__.py
@@ -1,14 +1,0 @@
-# default
-import asyncio
-
-# local
-from base import bot
-
-from .static_channels import static_channels
-
-
-@bot.listen()
-async def on_ready():
-    await asyncio.sleep(10)
-    static_channels.start()
-    print("1.Schedule ready")
